+++ src/org/starloco/locos/entity/mount/Mount.java
@@ -135,7 +135,7 @@ public class Mount {
 	
 	public Mount(int id, int color, int sexe, int amour, int endurance, int level, long exp, String name, int fatigue,
 			int energy, int reproduction, int maturity, int state, String objects, String ancestors, String capacitys, int size,
-			int cellId, short mapId, int owner, int orientation, long fecundatedHour, int couple, int savage) {
+			int cellId, int mapId, int owner, int orientation, long fecundatedHour, int couple, int savage) {
 		this.id = id;
 		this.color = color;
 		this.sex = sexe;
